+++ 3Sisters-Engine/inc/soundSystems/types/sound_buffer.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#ifndef SOUND_BUFFER_HPP
+#define SOUND_BUFFER_HPP
+
+// type alias to refer to a sound buffer
+using SoundBuffer = unsigned int;
+
+#endif
