+++ test/inc/test_window.hpp
@@ -4,10 +4,10 @@
 #define TEST_WINDOW_HPP
 
 #include <cameras/ortho_camera.hpp>
-#include <window/sisters_glfw_window.hpp>
-#include "input/sisters_glfw_gamepad.hpp"
+#include <window/sisters_sdl_window.hpp>
+#include <input/sisters_sdl_gamepad.hpp>
 
-using namespace GLFW;
+using namespace SDL;
 
 class TestWindow : public Window {
     private:
